feat: Main layout added and styled-system package added
