+++ react-ui-test/package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.5",
+    "styled-system": "^5.1.5",
     "typescript": "^4.8.3",
     "web-vitals": "^2.1.4"
   },
