+++ react-ui-test/src/components/Layout/Container/Container.tsx
@@ -1,0 +1,11 @@
+import { ContainerWrapper } from './Container.styles'
+
+interface ContainerTypes {
+  children: React.ReactNode
+}
+
+const Container = ({ children }: ContainerTypes) => {
+  return <ContainerWrapper>{children}</ContainerWrapper>
+}
+
+export default Container
