+++ react-ui-test/src/components/Layout/PageLayout/PageLayout.tsx
@@ -1,0 +1,11 @@
+import Footer from '../../UI/Footer/Footer';
+
+const PageLayout = () => {
+    return(
+        <div>
+            <Footer/>
+        </div>
+    )
+}
+
+export default PageLayout;
