+++ react-ui-test/src/components/UI/Footer/Footer.styles.ts
@@ -2,6 +2,4 @@ import styled from 'styled-components';
 
 export const FooterWrapper = styled.div`
     width: 100%;
-    height: 200px;
-    background-color: ${({theme}) => theme.colors.colorBlack};
 `
