--- react-ui-test/src/styles/theme.js
@@ -1,19 +1,0 @@
-const theme = {
-    colors: {
-        colorYellowNegative: 'rgb(249, 173, 29)',
-        colorGreenPositive: 'rgb(60, 187, 180)',
-        colorDarkBackground: 'rgba(0, 0, 0, .4)',
-        colorDarkerBackground: 'rgba(0, 0, 0, .6)',
-        colorDarkerGray: 'rgba(51, 51, 51, 1)',
-        colorDarkGray: 'rgba(70, 70, 70, 1)',
-        colorLightGray: 'rgba(235, 235, 235, 1)',
-        colorLightBackground: 'rgba(255, 255, 255, .4)',
-        colorLighterBackground: 'rgba(255, 255, 255, .8)',
-        colorWhite: 'rgba(255, 255, 255, 1)',
-        colorBlack: 'rgba(0, 0, 0, 1)',
-    }
-}
-
-export default {
-    ...theme
-}
